+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # @project-version@ (@build-time@)
 
-* Fixed a bug where the name of the target unit of a cast was not shown (a side-effect of the v13.0.19 bug fix)  [Comment #8286].
-* Fixed nameplates doubling in size after a reload or login when UI scale is not 100% and "Ignore UI Scale" is disabled [Comment #8240].
+* Fixed castbar color, shield, and overlay not updating correctly when a cast changes to interruptible or non-interruptible.